chore: restructure files and folders
