+++ src/App.jsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React from 'react';
 import { Routes, Route } from 'react-router-dom';
 import Home from './pages/Home';
 import Events from './pages/Events';
@@ -6,23 +6,32 @@ import About from './pages/About';
 import Contact from './pages/Contact';
 import Pricing from './pages/Pricing';
 import Help from './pages/Help';
-import Navbar from "./components/Navbar";
+import LandingLayout from "./components/LandingLayout";
+import AuthLayout from "./components/AuthLayout";
+import SignInForm from "./pages/auth/SignIn";
+import SignUpForm from "./pages/auth/SignUp";
 
 function App() {
   return (
-  <>
-  <Navbar />
-          <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/events" element={<Events />} />
-            <Route path="/about" element={<About />} />
-            <Route path="/contact" element={<Contact />} />
-            <Route path="/pricing" element={<Pricing />} />
-            <Route path="/help" element={<Help />} />
-          </Routes>
+    <>
+    <Routes>
+    {/* Landing pages with Navbar */}
+    <Route element={<LandingLayout />}>
+      <Route path="/" element={<Home />} />
+      <Route path="/events" element={<Events />} />
+      <Route path="/about" element={<About />} />
+      <Route path="/contact" element={<Contact />} />
+      <Route path="/pricing" element={<Pricing />} />
+      <Route path="/help" element={<Help />} />
+    </Route>
 
-      </>
-    
+    {/* Auth pages without Navbar */}
+    <Route element={<AuthLayout />}>
+      <Route path="/signin" element={<SignInForm />} />
+      <Route path="/signup" element={<SignUpForm />} />
+    </Route>
+  </Routes>
+  </>
   );
 }
