+++ src/components/AuthLayout.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Outlet } from "react-router-dom";
+
+function AuthLayout() {
+  return (
+    <div className="auth-container">
+      <Outlet />
+    </div>
+  );
+}
+
+export default AuthLayout;
