+++ src/components/Footer.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Footer() {
+    return (
+      <nav className="bg-white max-w-4xl mx-auto shadow-md">
+        <p>Footer</p>
+      </nav>
+    );
+  }
+  
+  export default Footer;
