+++ src/pages/Home.jsx
@@ -1,4 +1,6 @@
-import React from 'react';
+// import React from 'react';
+
+import React from "react";
 
 // import SignUpForm from '../components/forms/SignUpForm';
 // import SignInForm from '../components/forms/SignInForm';
