+++ src/pages/auth/SignIn.jsx
@@ -2,21 +2,18 @@
 import React from 'react';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
-import {
-  signUpSchema,
-  // SignUpFormData,
-} from '../../formValidators/authValidators';
+import { signInSchema } from '../../utils/authValidators';
 
-// NOTE: This is the SignUpForm component
+// NOTE: This is the SignInForm component
 // It should be further designed and styled as per the required UI design, this is just a basic implementation
 // to show how the form works with the created validations
-export default function SignUpForm() {
+export default function SignInForm() {
   const {
     register,
     handleSubmit,
     formState: { errors },
   } = useForm({
-    resolver: yupResolver(signUpSchema),
+    resolver: yupResolver(signInSchema),
   });
 
   const onSubmit = (data) => {
@@ -29,39 +26,6 @@ export default function SignUpForm() {
       onSubmit={handleSubmit(onSubmit)}
       className='flex flex-col items-center gap-4 '
     >
-      <div className='flex flex-col gap-2'>
-        <input
-          {...register('firstName')}
-          placeholder='First Name'
-          className='outline-none border-2 border-gray-300 rounded-md p-2 w-[500px]'
-        />
-        {errors.firstName && (
-          <span className='text-red-500'>{errors.firstName.message}</span>
-        )}
-      </div>
-
-      <div className='flex flex-col gap-2'>
-        <input
-          {...register('lastName')}
-          placeholder='Last Name'
-          className='outline-none border-2 border-gray-300 rounded-md p-2 w-[500px]'
-        />
-        {errors.lastName && (
-          <span className='text-red-500'>{errors.lastName.message}</span>
-        )}
-      </div>
-
-      <div className='flex flex-col gap-2'>
-        <input
-          {...register('username')}
-          placeholder='Username'
-          className='outline-none border-2 border-gray-300 rounded-md p-2 w-[500px]'
-        />
-        {errors.username && (
-          <span className='text-red-500'>{errors.username.message}</span>
-        )}
-      </div>
-
       <div className='flex flex-col gap-2'>
         <input
           {...register('email')}
@@ -87,7 +51,7 @@ export default function SignUpForm() {
       </div>
 
       <button type='submit' className='bg-blue-500 text-white p-2 rounded-md'>
-        Sign Up
+        Sign In
       </button>
     </form>
   );
